pyproject.toml
[project]
name = "retro-rag-assistant"
version = "0.1.0"
description = "RAG-based retrospective analysis service for moalog-platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.34",
    "asyncpg>=0.30",
    "aiomysql>=0.2",
    "redis[hiredis]>=5.0",
    "pgvector>=0.3",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "openai>=1.50",
    "langchain-text-splitters>=0.3",
    "langgraph>=0.2",
    "prometheus-client>=0.21",
    "tiktoken>=0.8",
    "httpx>=0.28",
]

[project.scripts]
retro-rag = "app.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
